pyproject.toml
[tool.poetry]
name = "schug"
version = "1.5"
description = "Keep track of genes, transcripts and exons from different sources"
authors = ["moonso <mans.magnusson@scilifelab.se>"]
readme = "README.md"

[tool.poetry.scripts]
schug = "schug.cli.base:cli"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
typer = "^0.4.0"
aiofiles = "^23.1.0"
httpx = "^0.23.3"
requests = "^2.28.2"
pytest-mock = "^3.10.0"
gunicorn = "^20.1.0"
importlib-resources = "^5.12.0"
uvicorn = "^0.21.1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"

[tool.poetry.dev-dependencies]
mkdocs = "^1.2.2"
mkdocs-material = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
